Move bitcoin price fetching to MutinyWallet
